LPS-36216 Replace getXYByUuidAndGroupId methods to fetchXYByUuidAndGroupId now that we have them for all models

+++ portal-impl/src/com/liferay/portlet/bookmarks/lar/BookmarksFolderStagedModelDataHandler.java
@@ -45,10 +45,12 @@ public class BookmarksFolderStagedModelDataHandler
 		throws PortalException, SystemException {
 
 		BookmarksFolder folder =
-			BookmarksFolderLocalServiceUtil.getBookmarksFolderByUuidAndGroupId(
-				uuid, groupId);
+			BookmarksFolderLocalServiceUtil.
+				fetchBookmarksFolderByUuidAndGroupId(uuid, groupId);
 
-		BookmarksFolderLocalServiceUtil.deleteFolder(folder);
+		if (folder != null) {
+			BookmarksFolderLocalServiceUtil.deleteFolder(folder);
+		}
 	}
 
 	@Override
